requirements.txt
beautifulsoup4==4.12.2
dateparser==1.1.8
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1